// src/app/claim-processing/manual-claim-processing/process-claim-no-pay/model/rpdmaa1.model.ts
import {ScreenChargeLines} from './screen-charge-lines.model';
import {ScreenIcdCodes} from './screen-icd-codes.model';
/**
 * Model class RPDMAA1
 * Path: screenbean/procclmnopaychrg
 * Model: com::uhc::aarp::fox::domain::screenbean::procclmnopaychrg::RPDMAA1
 */
export class RPDMAA1 {
  ma1cmnd = '';
  ma1mem1 = '';
  ma1mem2 = '';
  ma1mem3 = '';
  ma1cno1 = '';
  ma1cno2 = '';
  ma1cno3 = '';
  ma1cno4 = '';
  ma1cno5 = '';
  ma1cno6 = '';
  ma1com1 = '';
  ma1com2 = '';
  ma1name = '';
  ma1pln1 = '';
  ma1eff1 = '';
  ma1trm1 = '';
  ma1memn = '';
  ma1assc = '';
  ma1insr = '';
  ma1pln2 = '';
  ma1eff2 = '';
  ma1trm2 = '';
  ma1dob = '';
  ma1pln3 = '';
  ma1eff3 = '';
  ma1trm3 = '';
  ma1icd1 = '';
  ma1icd2 = '';
  ma1icd3 = '';
  ma1pln4 = '';
  ma1eff4 = '';
  ma1trm4 = '';
  ma1pla1 = '';
  ma1pro1 = '';
  ma1tos1 = '';
  ma1sr11 = '';
  ma1sr21 = '';
  ma1crg1 = '';
  ma1cpt1 = '';
  ma1pla2 = '';
  ma1pro2 = '';
  ma1tos2 = '';
  ma1sr12 = '';
  ma1sr22 = '';
  ma1crg2 = '';
  ma1cpt2 = '';
  ma1pla3 = '';
  ma1pro3 = '';
  ma1tos3 = '';
  ma1sr13 = '';
  ma1sr23 = '';
  ma1crg3 = '';
  ma1cpt3 = '';
  ma1pla4 = '';
  ma1pro4 = '';
  ma1tos4 = '';
  ma1sr14 = '';
  ma1sr24 = '';
  ma1crg4 = '';
  ma1cpt4 = '';
  ma1pla5 = '';
  ma1pro5 = '';
  ma1tos5 = '';
  ma1sr15 = '';
  ma1sr25 = '';
  ma1crg5 = '';
  ma1cpt5 = '';
  ma1pla6 = '';
  ma1pro6 = '';
  ma1tos6 = '';
  ma1sr16 = '';
  ma1sr26 = '';
  ma1crg6 = '';
  ma1cpt6 = '';
  ma1pla7 = '';
  ma1pro7 = '';
  ma1tos7 = '';
  ma1sr17 = '';
  ma1sr27 = '';
  ma1crg7 = '';
  ma1cpt7 = '';
  ma1pla8 = '';
  ma1pro8 = '';
  ma1tos8 = '';
  ma1sr18 = '';
  ma1sr28 = '';
  ma1crg8 = '';
  ma1cpt8 = '';
  ma1pla9 = '';
  ma1pro9 = '';
  ma1tos9 = '';
  ma1sr19 = '';
  ma1sr29 = '';
  ma1crg9 = '';
  ma1cpt9 = '';
  ma1plaa = '';
  ma1proa = '';
  ma1tosa = '';
  ma1sr1a = '';
  ma1sr2a = '';
  ma1crga = '';
  ma1cpta = '';
  ma1plab = '';
  ma1prob = '';
  ma1tosb = '';
  ma1sr1b = '';
  ma1sr2b = '';
  ma1cptb = '';
  ma1crgb = '';
  ma1plac = '';
  ma1proc = '';
  ma1tosc = '';
  ma1sr1c = '';
  ma1sr2c = '';
  ma1crgc = '';
  ma1cptc = '';
  ma1plad = '';
  ma1prod = '';
  ma1tosd = '';
  ma1sr1d = '';
  ma1sr2d = '';
  ma1crgd = '';
  ma1cptd = '';
  ma1err1 = '';
  ma1err2 = '';
  mapCommandLine = '';
  screenErr1 = '';
  screenErr2 = '';
  screenIcdCodes: ScreenIcdCodes[] = [];
  screenChargeLines: ScreenChargeLines[] = [];
}